requirements.txt
alpaca-trade-api==0.51.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
idna==2.10
install==1.3.4
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.0
jupyter-client==6.1.7
jupyter-core==4.7.0
numpy==1.19.4
pandas==1.2.0
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
Pygments==2.7.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
pywin32==300
pyzmq==20.0.0
requests==2.25.1
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
ta==0.4.5
threadpoolctl==2.1.0
tornado==6.1
-e git+https://github.com/DMinghao/Stock_Trading_Bot.git@1974d394435b8df21aca215869de94937383ab9b#egg=tradingBot
tradingview-ta==3.1.2
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
